+++ .idea/misc.xml
@@ -32,9 +32,10 @@
         <entry key="..\:/Users/COP412L/AndroidStudioProjects/ficha/app/src/main/res/layout/activity_perfil.xml" value="0.33" />
         <entry key="..\:/Users/COP412L/AndroidStudioProjects/ficha/app/src/main/res/layout/fragment_info_pasajero.xml" value="0.358695652173913" />
         <entry key="..\:/Users/COP412L/AndroidStudioProjects/ficha2/app/src/main/res/layout/activity_crear_cuenta2.xml" value="0.358695652173913" />
-        <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_destino.xml" value="0.22" />
+        <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_destino.xml" value="0.12" />
         <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_destino1.xml" value="0.16847826086956522" />
         <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_destino_places.xml" value="0.18885869565217392" />
+        <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_google_maps_api.xml" value="0.17119565217391305" />
         <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_info_conductor.xml" value="0.14221014492753623" />
         <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_info_pasajero.xml" value="0.19927536231884058" />
         <entry key="..\:/Users/DUVAN/AndroidStudioProjects/Ficha_1/app/src/main/res/layout/activity_login.xml" value="0.16847826086956522" />
